+++ doc/changelog.rst
@@ -12,6 +12,8 @@ What's New
 
   * Fixes things not appearing in WARC files. This regression was introduced in 0.26 where PhantomJS's disk cache was enabled. It is now disabled again.
   * Fixes HTTPS URL rewriting where relative URLs were not properly rewritten.
+  * Fixes unwanted ``Accept-Language`` header picked up from environment. The value has been overridden to ``*``.
+  * Fixes ``--header`` options left out in requests.
 
 * API: